update_test

+++ ProblemSolving.Test/FibonacciSequence_Test.cs
@@ -1,5 +1,4 @@
 ﻿using ProblemSolving.Core;
-using System.Numerics;
 
 namespace ProblemSolving.Test
 {
@@ -8,7 +7,7 @@ namespace ProblemSolving.Test
         [Theory]
         [InlineData(0, 1, 6, 27)]
         [InlineData(0, 1, 5, 5)]
-        public void GetFibonacci(BigInteger t1, BigInteger t2, int n, BigInteger result)
+        public void GetFibonacci(int t1, int t2, int n, int result)
         {
             var res = FibonacciSequence.FibonacciModified(t1, t2, n);
             Assert.Equal(result, res);
